SMACKBench - Add support for identifier field on results webpage
This commit uses the benchexec rundefinition_name field from
benchexec xml files to add a description after the set name
(the rundefinition_name field was previously simply the set
name).

This will allow runs to be distinguishable from the results webpage
by something other than simply remembering which runs were which.

A command line option, -d --description, has been added to SMACKBench.py
which is appended to the set name in the rundefinition_name field.

I chose to do this way (using rundefinition_name field) to avoid
having to use custom schema on the benchexec in/out xml files.
An added benefit of this choice is that the description also shows up
in the resulting tables, so the runs can also be distinguished in the
table view.